consumer/requirements.txt
uvicorn==0.14.0
fastapi==0.68.0
requests==2.26.0
scikit-learn==0.24.2
pyspark-stubs==2.3.0
